ai/ai_logic.py
# ai/ai_logic.py
import io
import random

from PIL import Image


def generate_ai_text():
    # Placeholder logic for AI text creation; replace with actual AI logic or more sophisticated generation
    responses = [
        "A sunny day at the beach.",
        "A cat chasing a mouse.",
        "An astronaut floating in space.",
        "A dragon flying over a castle.",
        "A robot making coffee.",
        "A magical forest with unicorns.",
        "A pirate ship battling a sea monster."
    ]
    return random.choice(responses)

def generate_ai_guess(drawing_path):
    # Placeholder logic for AI guessing based on a drawing
    # In a real implementation, integrate with image recognition or AI image description models
    # Here, we'll return a random guess
    guesses = [
        "A cat chasing a ball.",
        "A spaceship landing on Mars.",
        "A wizard casting a spell.",
        "A car racing on a track.",
        "A tree with colorful leaves.",
        "A dog playing fetch.",
        "A house by the lake."
    ]
    return random.choice(guesses)

def generate_ai_drawing(prompt):
    # Placeholder: Return empty bytes or integrate with an image generation API
    # For demonstration, we'll return empty bytes to represent an image
    # Replace this with actual image generation logic if desired
    img = Image.open("temp_ai_img.jpg")
    # Create a BytesIO object to save the image in memory
    byte_io = io.BytesIO()
    img.save(byte_io, format='PNG')
    
    # Retrieve the byte data
    byte_data = byte_io.getvalue()
    return byte_data